Fawad 4 pm 20/9/2019

+++ app/src/main/java/com/app/leafy/model/ProductOrder.java
@@ -12,7 +12,7 @@ public class ProductOrder implements Serializable {
     public String phone;
     public String comment;
     public String status;
-    public Double total_fees;
+    public Integer total_fees;
     public Double tax;
     public String serial;
     public Long created_at = System.currentTimeMillis();
